chore: Project initialization (#5)
* chore(functions): Initialize Firebase Functions and Emulators

Configured Functions to use the "media-metadata-manager" codebase tag as defined in [this Firebase doc](https://firebase.google.com/docs/functions/organize-functions#managing_multiple_repositories) to prevent conflicts between functions from multiple hackathon projects on the same Firebase Project.

Configured emulators for Auth, Functions, Firestore, and Storage, as well as the Emulator UI. This should allow for entirely offline development and testing.

Did not configure Firestore or Storage rules and/or indices because they don't play nice with other repos, and would overwrite the settings set anywhere else.

Working on #3.

* chore(website): Initialize the Front End using create-react-app

Working on #1.

* chore(website): Install MUI Packages

Installed the main MUI Core components, as well as their icon library. Also going to try using the Emotion library for styling.

Working on #1.

* chore(website): Installed the Roboto Font via CDN

Based on the instructions from https://mui.com/material-ui/react-typography/#general.

Completes #1.

* chore: Configure Firestore and Firebase Storage

Required to run those in the emulator.

Working on #3.

* chore(website): Configure Firebase SDK for the Website

Complete basic initialization of the different components of the Firebase SDK on the website,

Completes #3.

* chore(functions): Configure the Basic Express App

Completes #4.

* chore(functions): Configure GraphQL Endpoint

Create a basic GraphQL "Hello World" service to act as the API Interface.

Working on #2.

* chore(website): Configure GraphQL Client Library

Note: needed to pull out some of the internal code from the Firebase Functions library to figure out what URL to use for the GraphQL Client, but this was easier than trying to modify the Firebase Functions SDK to handle GraphQL queries.

Completes #2.