feat(functions/graphql): Define Status and DateTime Resolvers
Logic is currently for static data; to be replaced with the actual Firestore logic later.

Note: I had to switch from an actual enum to a set of strings for the `StatusOptions` TypeScript` type because the `@types` folder is not copied to the final build, and the enum would have been required in the function to run.

Closes #15.